refactored tables and started created tests
